+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 --------------
 
 - Fix FIR diagnostics rendering in the IDE. Note this only works in the K2 Kotlin IDE plugin + setting the IntelliJ `kotlin.k2.only.bundled.compiler.plugins.enabled` registry key to `false`.
+- When custom annotations are defined, report those names in FIR error messages.
 - Build against Gradle `8.12`.
 - Only report errors in FIR now. Removes the `validateIr` plugin option.
 - No longer support K1.